+++ app/models/document.rb
@@ -16,10 +16,35 @@
 #
 
 class Document < ActiveRecord::Base
+  track_who_does_it
+  attr_accessible :event_id, :filename, :date, :type, :title, :is_public,
+    :language_id, :citations_attributes
   belongs_to :event
   belongs_to :language
+  has_many :citations, class_name: 'DocumentCitation'
   validates :title, presence: true
   validates :date, presence: true
-  validates :is_public, presence: true
   # TODO validates inclusion of type in available types
+  accepts_nested_attributes_for :citations, :allow_destroy => true,
+    :reject_if => proc { |attributes|
+    attributes['stringy_taxon_concept_ids'].blank? && (
+      attributes['geo_entity_ids'].blank? || attributes['geo_entity_ids'].reject(&:blank?).empty?
+    )
+  }
+  mount_uploader :filename, DocumentFileUploader
+
+  before_validation :set_title
+
+  def self.display_name; 'Document'; end
+
+  def set_title
+    if title.blank? && filename_changed?
+      self.title = filename.file.filename.sub(/.\w+$/, '').humanize
+    end
+  end
+
+  def date_formatted
+    date && date.strftime("%d/%m/%Y")
+  end
+
 end