+++ config/initializers/carrierwave.rb
@@ -1,0 +1,7 @@
+# config/initializers/carrierwave.rb
+if Rails.env.test? || Rails.env.cucumber?
+  CarrierWave.configure do |config|
+    config.storage = :file
+    config.enable_processing = false
+  end
+end