+++ db/migrate/20140812133548_document_is_public_should_default_to_false.rb
@@ -1,0 +1,5 @@
+class DocumentIsPublicShouldDefaultToFalse < ActiveRecord::Migration
+  def change
+    change_column :documents, :is_public, :boolean, null: false, default: false
+  end
+end